requirements.txt
accessible-pygments==0.0.5
affine==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.1.0
aiosignal==1.3.1
alabaster==1.0.0
albucore==0.0.19
albumentations==1.4.20
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
autopep8==2.3.1
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
commonmark==0.9.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
deepforest==1.4.1
defusedxml==0.7.1
docformatter==1.7.5
docutils==0.21.2
eval_type_backport==0.2.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
geopandas==1.0.1
huggingface-hub==0.26.1
idna==3.10
imagesize==1.4.1
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
lightning-utilities==0.11.8
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nbqa==1.9.0
networkx==3.4.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
progressbar2==4.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pydata-sphinx-theme==0.16.0
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-utils==3.9.0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.1
recommonmark==0.7.1
requests==2.32.3
Rtree==1.3.0
safetensors==0.4.5
scipy==1.14.1
shapely==2.0.6
six==1.16.0
slidingwindow==0.0.14
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
stringzilla==3.10.5
supervision==0.24.0
sympy==1.13.3
timm==1.0.11
tokenize_rt==6.1.0
tomli==2.0.2
torch==2.4.1
torchmetrics==1.5.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
untokenize==0.1.1
urllib3==2.2.3
wcwidth==0.2.13
xmltodict==0.14.2
yarl==1.16.0
